requirements.txt
aiogram==3.15.0
aiosqlite==0.20.0
paramiko==3.5.0
psutil==6.1.0
apscheduler==3.10.4
python-dotenv==1.0.1
aiohttp==3.11.0
